Separate splicing and refresh into separate plugins
