+++ example/build.sbt
@@ -1,5 +1,10 @@
 enablePlugins(ScalaJSPlugin)
+
+// dynamic page reloading
 enablePlugins(WorkbenchPlugin)
+
+// (experimental feature) in-place code update with state preservation
+// enablePlugins(WorkbenchSplicePlugin) // disable WorkbenchPlugin when activating
 
 name := "Example"
 
@@ -10,6 +15,3 @@ version := "0.1-SNAPSHOT"
 libraryDependencies ++= Seq(
   "org.scala-js" %%% "scalajs-dom" % "0.9.1"
 )
-
-// (experimental feature)
-spliceBrowsers <<= spliceBrowsers.triggeredBy(fastOptJS in Compile)
